Add files via upload
This project focuses on predicting Hepatitis outcomes using logistic regression on a small dataset. With only a limited number of samples, the model achieved an accuracy of 89.47%, demonstrating strong performance in identifying patients who survive or succumb to Hepatitis. Despite the dataset's size, the model exhibited a sensitivity of 100%, correctly predicting all fatalities. The specificity reached 87.50%, showcasing its ability to accurately identify survivors. Results should be interpreted cautiously due to the dataset's small scale, but the promising metrics highlight the model's potential for predicting Hepatitis outcomes.